Separate nav into its own page
